Refactor Maven configuration to establish correct parent-child inheritance.
Centralize dependency version management in the root POM.
Remove redundant AOP modules and auto-generated test files.
Perform global code reformatting: fixed indentation and optimized imports.

+++ project/delivery-desktop/src/main/java/com/example/service/RestaurantService.java
@@ -5,6 +5,7 @@ import com.example.entity.restaurant.Dish;
 import com.example.entity.restaurant.Restaurant;
 import com.example.repository.RestaurantRepository;
 import jakarta.persistence.EntityManager;
+
 import java.util.List;
 
 public class RestaurantService {
@@ -65,7 +66,7 @@ public class RestaurantService {
             em.getTransaction().begin();
 
             RestaurantRepository repo = new RestaurantRepository(em);
-            repo.deleteDish((long) dish.getDishId());
+            repo.deleteDish(dish.getDishId());
 
             em.getTransaction().commit();
         } catch (Exception e) {
